refactor: extract GitCooldownDateResolver shared module
Deduplicates cooldown date resolution logic between pre-commit and
GitHub Actions into a common module. Both ecosystems now include
Dependabot::GitCooldownDateResolver and implement two bridge methods:
- cooldown_source_url: returns the git source URL
- cooldown_credentials: returns the credentials array

Shared methods: normalize_tag_name, resolve_candidate_date,
github_release_published_at, tag_creation_date, cached_github_releases.
